feat (app.js): date input validation for 'smart commands'; skip to next field upon 'Enter' keypress
Developer-named 'smart commands' are sequences of characters which trigger fast date input functionality.
While any of the month, day, or year fields in the date input are focused, if a user presses 'Enter' immediately following a case-insensitive 'D', 'M', or 'Y' keypress and an Arabic numeral keypress, a smart command is triggered that adjusts the input's value accordingly.
For example, given the current date is 2022-12-22 and the user presses the keys 'd', '3', and 'Enter' (in that order), the date input would populate with the value "2023-12-25".
Another example with the same date, the user presses the keys 'M', '5', and 'Enter', the date input would populate with the value "2023-5-22".